+++ common/src/main/java/net/spell_engine/rpg_series/RPGSeriesCore.java
@@ -16,7 +16,7 @@ public class RPGSeriesCore {
     public static final String NAMESPACE = "rpg_series";
 
     public static ConfigManager<LootConfig> lootEquipmentConfig = new ConfigManager<>
-            ("loot_equipment_v3", LootDefaults.itemLootConfig)
+            ("loot_equipment_v2", LootDefaults.itemLootConfig)
             .builder()
             .setDirectory(NAMESPACE)
             .sanitize(true)